requirements.txt
gpt-2-simple==0.8.1
discord.py==1.7.3
